Create 1.5 Basic string functions
/*
1. For character array = string declaration size will be +1 for null terminator

2. Size will be automatically understood by complier

3. Curly braces are not mandatory for character array

4. When printf("%d\n",  (str2[0]=='\0') if true than 1 will be print 
               
5. Always initialize your string 

6. strlen() is for string length (count without backslash zero) and sizeof()
for number of byte. this count with backslash zero
    
7. strlen() and strcpy() are function of <string.h> those are related to sting funtions
and <ctype.h> related to the function related to character functions   
               
8.We can't directly copy one string name to another with = sign and this sign is 
used for only initialization not for copying and strcpy(destination, source) is used for this

9. strcmp(firstone, seconeone) for compare when both are same (will be 0) otherwise
it will be 1. It compares alphabetically when first one is less then second one then 
answer will be -1. =(first one - second one)

10. strcat() is for add to string to another string
               
11. gets(str) can input string with spaces that scanf("%s",str) . puts(str) is to 
print string if you want to print 

12. isalpha(),isdigit(), isalnum(), islower(), isupper(), tolower('H'), toupper(). 
isspace().  all are 0 for false and true is otherwise. google ctype.h headerfiles to learn more
                
*/